refactor: RIP obsolete files and functionalities
